src/detr/model_DETR/detr/requirements_detr.txt
absl-py==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.1
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.10.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
huggingface-hub==0.21.2
idna==2.10
imageio==2.34.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lightning @ https://github.com/Lightning-AI/lightning/archive/refs/heads/master.zip#sha256=fde694a72ad093012f23e71b19432f947ee426688d8a92b4bd98554a4a2929aa
lightning-utilities==0.10.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.2.0
parso==0.8.3
Pillow==10.0.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.24.2
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycocotools==2.0.7
pyDeprecate==0.3.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytorch-lightning==1.7.7
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.21
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
supervision==0.18.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2024.2.12
timm==0.9.16
tokenizers==0.13.3
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchdata==0.5.1
torchmetrics==0.11.4
torchtext==0.14.1
torchvision==0.14.1+cu117
tqdm==4.66.2
traitlets==5.14.1
transformers==4.28.0
typing_extensions==4.10.0
tzdata==2023.4
urllib3==1.26.16
wcwidth==0.2.13
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.4
zipp==3.16.2
